+++ src/core/layers/tests/unit/group.spec.js
@@ -2,6 +2,7 @@ import {expect} from "chai";
 import sinon from "sinon";
 import WMSLayer from "../../wms";
 import WFSLayer from "../../wfs";
+import OAFLayer from "../../oaf";
 import GroupedLayers from "../../group";
 import mapCollection from "../../../../core/dataStorage/mapCollection.js";
 import store from "../../../../app-store";
@@ -9,6 +10,7 @@ import store from "../../../../app-store";
 describe("src/core/layers/group.js", () => {
     let wmsLayerAttributes,
         wfsLayerAttributes,
+        oafLayerAttributes,
         groupAttributes;
     const olLayer = {
         values_: {opacity: 1},
@@ -65,6 +67,16 @@ describe("src/core/layers/group.js", () => {
             featureNS: "http://www.deegree.org/app",
             featureType: "krankenhaeuser_hh"
         };
+        oafLayerAttributes = {
+            name: "oafTestLayer",
+            id: "id_oaf_layer",
+            url: "https://url.de",
+            typ: "OAF",
+            isChildLayer: true,
+            transparent: false,
+            featureNS: "http://www.deegree.org/app",
+            featureType: "krankenhaeuser_hh"
+        };
         groupAttributes = {
             name: "groupTestLayer",
             id: "id_group",
@@ -87,19 +99,22 @@ describe("src/core/layers/group.js", () => {
 
     it("createLayer shall create an ol.Group with olLayer in layerSource", function () {
         groupAttributes.children.push(wfsLayerAttributes);
+        groupAttributes.children.push(oafLayerAttributes);
         const groupUpdateSourceTrigger = sinon.spy(GroupedLayers.prototype, "updateSource"),
             groupCreateLegendTrigger = sinon.spy(GroupedLayers.prototype, "createLegend"),
             groupLayer = new GroupedLayers(groupAttributes),
             childLayer = groupLayer.get("layerSource"),
             wmsChildLayer = childLayer.find(layer => layer.get("typ") === "WMS"),
-            wfsChildLayer = childLayer.find(layer => layer.get("typ") === "WFS");
+            wfsChildLayer = childLayer.find(layer => layer.get("typ") === "WFS"),
+            oafChildLayer = childLayer.find(layer => layer.get("typ") === "OAF");
 
         expect(groupLayer).not.to.be.undefined;
         expect(wmsChildLayer).not.to.be.undefined;
         expect(wfsChildLayer).not.to.be.undefined;
+        expect(oafChildLayer).not.to.be.undefined;
         expect(groupLayer.get("layer").values_.visible).to.be.false;
-        expect(groupLayer.get("layer").getLayers().getLength()).to.be.equals(2);
-        expect(childLayer).to.be.an("array").with.lengthOf(2);
+        expect(groupLayer.get("layer").getLayers().getLength()).to.be.equals(3);
+        expect(childLayer).to.be.an("array").with.lengthOf(3);
         expect(wmsChildLayer.get("id")).to.be.equals(wmsLayerAttributes.id);
         expect(wmsChildLayer.get("layer")).not.to.be.undefined;
         expect(wmsChildLayer.get("typ")).to.be.equals(wmsLayerAttributes.typ);
@@ -107,6 +122,9 @@ describe("src/core/layers/group.js", () => {
         expect(wfsChildLayer.get("id")).to.be.equals(wfsLayerAttributes.id);
         expect(wfsChildLayer.get("typ")).to.be.equals(wfsLayerAttributes.typ);
         expect(wfsChildLayer).to.be.an.instanceof(WFSLayer);
+        expect(oafChildLayer.get("id")).to.be.equals(oafLayerAttributes.id);
+        expect(oafChildLayer.get("typ")).to.be.equals(oafLayerAttributes.typ);
+        expect(oafChildLayer).to.be.an.instanceof(OAFLayer);
         expect(groupUpdateSourceTrigger.calledOnce).to.be.false;
         expect(groupCreateLegendTrigger.calledOnce).to.be.true;
     });
@@ -121,15 +139,18 @@ describe("src/core/layers/group.js", () => {
     it("createLayer with isVisibleInMap=true shall call update source for each layer", function () {
         const groupUpdateSourceTrigger = sinon.spy(GroupedLayers.prototype, "updateSource"),
             wmsUpdateSourceTrigger = sinon.spy(WMSLayer.prototype, "updateSource"),
-            wfsUpdateSourceTrigger = sinon.spy(WFSLayer.prototype, "updateSource");
+            wfsUpdateSourceTrigger = sinon.spy(WFSLayer.prototype, "updateSource"),
+            oafUpdateSourceTrigger = sinon.spy(OAFLayer.prototype, "updateSource");
 
         groupAttributes.children.push(wfsLayerAttributes);
+        groupAttributes.children.push(oafLayerAttributes);
         groupAttributes.isVisibleInMap = true;
         new GroupedLayers(groupAttributes);
 
         expect(groupUpdateSourceTrigger.calledOnce).to.be.true;
         expect(wmsUpdateSourceTrigger.calledOnce).to.be.true;
         expect(wfsUpdateSourceTrigger.calledOnce).to.be.true;
+        expect(oafUpdateSourceTrigger.calledOnce).to.be.true;
     });
     it("showLayerInformation shall dispatch", function () {
         wmsLayerAttributes.datasets = [{